added data loading

+++ app/src/main/java/com/example/kinohype/MainActivity.java
@@ -1,10 +1,13 @@
 package com.example.kinohype;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
+import androidx.loader.app.LoaderManager;
+import androidx.loader.content.Loader;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
@@ -26,18 +29,27 @@ import com.example.kinohype.data.ViewModel;
 
 import org.json.JSONObject;
 
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 
-public class MainActivity extends AppCompatActivity {
+public class MainActivity extends AppCompatActivity implements LoaderManager.LoaderCallbacks<JSONObject> {
 
     private Button buttonPop;
     private Button buttonTop;
     private Button buttonBes;
     private AdapterForMovie adapterForMovie;
     private RecyclerView recyclerViewImages;
+    //номер страницы
+    private static int page = 1;
+    //метод сортировки
+    private static int mSort;
+    private static boolean isLoading = false;
 
     private ViewModel viewModel;
+    //идентификатор загрузчика
+    private static final int loaderid = 932;
+    private LoaderManager loaderManager;
 
     //переопределяем метод с меню
     @Override
@@ -70,6 +82,8 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        //отвечает за загрузки в приложении
+        loaderManager = LoaderManager.getInstance(this);
         viewModel = ViewModelProviders.of(this).get(ViewModel.class);
         //кнопки для переключения подборки кино
         buttonTop = findViewById(R.id.buttonTop);
@@ -93,33 +107,39 @@ public class MainActivity extends AppCompatActivity {
 
             @Override
             public void onClick(View v) {
+                page = 1;
+                mSort = Internet.POP;
                 //меняем цвет текста на кнопках
                 buttonPop.setTextColor(getResources().getColor(R.color.chooseMovieType));
                 buttonTop.setTextColor(getResources().getColor(R.color.defaultColor));
                 buttonBes.setTextColor(getResources().getColor(R.color.defaultColor));
-                downloadData(Internet.POP, 1);
+                downloadData(Internet.POP, page);
             }
         });
 
         buttonBes.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                page = 1;
+                mSort = Internet.BES;
                 //меняем цвет текста на кнопках
                 buttonPop.setTextColor(getResources().getColor(R.color.defaultColor));
                 buttonTop.setTextColor(getResources().getColor(R.color.defaultColor));
                 buttonBes.setTextColor(getResources().getColor(R.color.chooseMovieType));
-                downloadData(Internet.BES, 1);
+                downloadData(Internet.BES, page);
             }
         });
 
         buttonTop.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                page = 1;
+                mSort = Internet.TOP;
                 //меняем цвет текста на кнопках
                 buttonPop.setTextColor(getResources().getColor(R.color.defaultColor));
                 buttonTop.setTextColor(getResources().getColor(R.color.chooseMovieType));
                 buttonBes.setTextColor(getResources().getColor(R.color.defaultColor));
-                downloadData(Internet.TOP, 1);
+                downloadData(Internet.TOP, page);
             }
         });
         adapterForMovie.setOnImageMovieClickListener(new AdapterForMovie.OnImageMovieClickListener() {
@@ -134,9 +154,13 @@ public class MainActivity extends AppCompatActivity {
             }
         });
         adapterForMovie.setOnSetInTheEnd(new AdapterForMovie.OnSetInTheEnd() {
+
             @Override
             public void onSetEnd() {
                 //переопределяем метод
+                if(!isLoading) {
+                    downloadData(mSort, page);
+                }
 
             }
         });
@@ -145,21 +169,53 @@ public class MainActivity extends AppCompatActivity {
         moviesLiveData.observe(this, new Observer<List<Movie>>() {
             @Override
             public void onChanged(List<Movie> movies) {
-                adapterForMovie.setM(movies);
+                if(page == 1) adapterForMovie.setM(movies);
+
             }
         });
     }
     private void downloadData(int method, int page) {
-        JSONObject jsonObject = Internet.getJSONObjectfromInternet(method, page);
-        //получаем список фильмов
-        ArrayList<Movie> movies = JSONformat.getMovieJSON(jsonObject);
+       URL url = Internet.buildURL(method, page);
+       Bundle bundle = new Bundle();
+       bundle.putString("url", url.toString());
+       loaderManager.restartLoader(loaderid, bundle, this);
+    }
+
+    @NonNull
+    @Override
+    public Loader<JSONObject> onCreateLoader(int id, @Nullable Bundle args) {
+        Internet.JSONloader jsoNloader = new Internet.JSONloader(this, args);
+        jsoNloader.setOnStartLoadingListener(new Internet.JSONloader.OnStartLoadingListener() {
+            @Override
+            public void onStartLoading() {
+                isLoading = true;
+            }
+        });
+        return jsoNloader;
+    }
+
+    @Override
+    public void onLoadFinished(@NonNull Loader<JSONObject> loader, JSONObject data) {
+        ArrayList<Movie> movies = JSONformat.getMovieJSON(data);
         if(movies != null && !movies.isEmpty()) {
             //очищаем старые данные
-            viewModel.deleteAllMovies();
+            if(page == 1) {
+                viewModel.deleteAllMovies();
+                adapterForMovie.clearM();
+            }
             //вставляем новые данные в цикле
-            for(Movie movie : movies) {
+            for (Movie movie : movies) {
                 viewModel.insertMovie(movie);
             }
+            adapterForMovie.addM(movies);
+            page++;
         }
+        isLoading = false;
+        loaderManager.destroyLoader(loaderid);
+    }
+
+    @Override
+    public void onLoaderReset(@NonNull Loader<JSONObject> loader) {
+
     }
 }
